fix: 트랙잭션 미적용 버그 수정
