+++ src/main/java/com/groom/sumbisori/domain/user/service/CustomOAuth2UserService.java
@@ -19,6 +19,7 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class CustomOAuth2UserService extends DefaultOAuth2UserService {
     private final UserRepository userRepository;
+    private final UserService userService;
 
     @Override
     public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException {
@@ -29,26 +30,7 @@ public class CustomOAuth2UserService extends DefaultOAuth2UserService {
         }
         OAuth2Response oAuth2Response = new KaKaoResponse(oAuth2User.getAttributes());
 
-        User user = createAndUpdateMember(oAuth2Response, registrationId);
+        User user = userService.createAndUpdateMember(oAuth2Response, registrationId);
         return new CustomUserDetails(user, oAuth2User.getAttributes());
-    }
-
-    @Transactional
-    public User createAndUpdateMember(OAuth2Response oAuth2Response, String providerType) {
-        return userRepository.findByProviderTypeAndProviderId(providerType, oAuth2Response.getProviderId())
-                .map(existingUser -> updateMember(existingUser, oAuth2Response))
-                .orElseGet(() -> registerNewMember(oAuth2Response));
-    }
-
-    private User updateMember(User existingUser, OAuth2Response oAuth2Response) {
-        existingUser.update(oAuth2Response.getEmail(), oAuth2Response.getNickname(),
-                oAuth2Response.getProfileImage());
-        return existingUser;
-    }
-
-    private User registerNewMember(OAuth2Response oAuth2Response) {
-        log.info("소셜로그인으로 처음 로그인(강제 회원가입): {}", oAuth2Response.getProvider());
-        User user = oAuth2Response.toEntity();
-        return userRepository.save(user);
     }
 }
