+++ src/main/java/com/groom/sumbisori/domain/user/service/UserService.java
@@ -1,6 +1,7 @@
 package com.groom.sumbisori.domain.user.service;
 
 import com.groom.sumbisori.domain.collection.repository.SeafoodCollectionQueryRepository;
+import com.groom.sumbisori.domain.user.dto.common.OAuth2Response;
 import com.groom.sumbisori.domain.user.dto.response.UserProfile;
 import com.groom.sumbisori.domain.user.entity.User;
 import com.groom.sumbisori.domain.user.error.UserErrorCode;
@@ -8,10 +9,12 @@ import com.groom.sumbisori.domain.user.error.exception.UserException;
 import com.groom.sumbisori.domain.user.oauth2.KakaoClient;
 import com.groom.sumbisori.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
+@Slf4j
 @RequiredArgsConstructor
 public class UserService {
     private final UserDeleteService userDeleteService;
@@ -36,5 +39,24 @@ public class UserService {
                 .orElseThrow(() -> new UserException(UserErrorCode.USER_NOT_FOUND));
         userDeleteService.delete(user);
         kakaoClient.unlink(user.getProviderId());
+    }
+
+    @Transactional
+    public User createAndUpdateMember(OAuth2Response oAuth2Response, String providerType) {
+        return userRepository.findByProviderTypeAndProviderId(providerType, oAuth2Response.getProviderId())
+                .map(existingUser -> updateMember(existingUser, oAuth2Response))
+                .orElseGet(() -> registerNewMember(oAuth2Response));
+    }
+
+    private User updateMember(User existingUser, OAuth2Response oAuth2Response) {
+        existingUser.update(oAuth2Response.getEmail(), oAuth2Response.getNickname(),
+                oAuth2Response.getProfileImage());
+        return existingUser;
+    }
+
+    private User registerNewMember(OAuth2Response oAuth2Response) {
+        log.info("소셜로그인으로 처음 로그인(강제 회원가입): {}", oAuth2Response.getProvider());
+        User user = oAuth2Response.toEntity();
+        return userRepository.save(user);
     }
 }
